requirements.txt
airtest==1.1.8
attrdict==2.0.1
attrs==20.3.0
Automat==20.2.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
decorator==4.4.2
Deprecated==1.2.11
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
facebook-wda==1.3.3
gitdb2==2.0.6
GitPython==2.1.15
hrpc==1.0.8
hyperlink==21.0.0
idna==2.7
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==2.6.0
lxml==4.6.2
MarkupSafe==1.1.1
mss==4.0.3
nose==1.3.7
numpy==1.18.5
opencv-contrib-python==4.5.1.48
parsel==1.6.0
Pillow==7.2.0
pocoui==1.0.79
Protego==0.1.16
py==1.10.0
pyaml==18.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyGithub==1.43.5
PyHamcrest==2.0.2
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-gitlab==1.8.0
python-xlib==0.29
pywinauto==0.6.3
PyYAML==3.13
queuelib==1.5.0
requests==2.20.1
retry==0.9.2
Scrapy==2.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
smmap==3.0.4
smmap2==3.0.1
svn==0.3.46
texttable==0.9.1
Twisted==20.3.0
urllib3==1.24.3
w3lib==1.22.0
websocket-client==0.48.0
wrapt==1.12.1
zope.interface==5.2.0